clockwork/clockwork_project/clockwork-src/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clockwork"
version = "0.1.0"
description = "Local-first repository intelligence and AI agent coordination system"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "gitpython>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]
graph = [
    "networkx>=3.0",
]
parse = [
    "tree-sitter>=0.21",
]

[project.scripts]
clockwork = "clockwork.cli.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["clockwork*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]
addopts = "-ra"

[tool.coverage.run]
source = ["clockwork"]
omit = ["tests/*"]
